% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCREPstack.R
\name{aggregateCREPstack}
\alias{aggregateCREPstack}
\title{aggregateCREPstack}
\usage{
aggregateCREPstack(r)
}
\arguments{
\item{r}{Raster stack from output from rasterCREP}

\item{fun}{function Default to sum}
}
\description{
aggregateCREPstack
}

