% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{API}
\alias{API}
\title{Air Pollutin Index}
\usage{
API(conc = conc, verbose = FALSE, cast.df = TRUE, out = "out",
  TZ = "Africa/Johannesburg")
}
\arguments{
\item{conc}{Raster that contains the concentrations.}

\item{verbose}{Logical that displays function messages if TRUE.}

\item{cast.df}{Logical that will cast the data frame if true.}

\item{out}{Character vector that contains the name of the casted
data frame to be assigned to the global environment.}

\item{TZ}{Character vector that contains the city}
}
\description{
Function
}

