\name{AQoffset-package}
\alias{AQoffset-package}
\alias{AQoffset}
\docType{package}
\title{
Makes calculation needed for air quality offsetting
}
\description{
Uses rasterised outputs of dispersion models or other air quality models and population data to calculate air quality impacts and the inpact of offsets
}
\details{
\tabular{ll}{
Package: \tab AQoffset\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-10\cr
License: \tab GPL \cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Christiaan.J. Pauw

Maintainer: Christiaan.J. Pauw
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
