% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_exceed.R
\name{bar_exceed}
\alias{bar_exceed}
\title{Bar Exceed}
\usage{
bar_exceed(z, cap = "", ttl = "", axn = TRUE, ces.ax = 0.9,
  yl = "Events: days x cell", xl = expression(paste(mu, plain(g/m)^3)), ...)
}
\arguments{
\item{z}{A raster or raster brick}

\item{cap}{Character vector. Contains the caption}

\item{ttl}{Character vector. Contains the title}

\item{axn}{Logical that draws the other axis (with lty = 0) and labels it if TRUE.}

\item{ces.ax}{Numerical used for axis annotation}

\item{yl}{A label for the y axis}

\item{xl}{A label for the x axis}

\item{...}{Arguments to be passed to/from other methods}
}
\description{
Makes a barplot of n exceedance object
}

