% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_implement_prop.R
\name{calc_implement_prop}
\alias{calc_implement_prop}
\title{calc_implement_prop}
\usage{
calc_implement_prop(PA = hh_24, MA = eskom_deficit, aveperiodlab = "24h",
  idpos = 1, polpos = 2, aveperiodpos = 3, cyclepos = 4, sep = "_")
}
\arguments{
\item{PA}{RasterStack}

\item{MA}{RasterStack}

\item{idpos}{Numeric referring to the ID position}

\item{polpos}{Numeric referring to the position of the pollutant.}

\item{aveperiodpos}{Numeric referring to the average period position.}

\item{cyclepos}{Numeric referring to the cycle position.}

\item{sep}{Character vector containing the seperator to be used.}
}
\description{
Bereken die proporsie van bron PA wat benodig word om bron MA af te set
}

