% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateGMMosaic.R
\name{consolidateGMMosaic}
\alias{consolidateGMMosaic}
\title{consolidateGMMosaic Function to download and stich together googleMaps 
images at a given resolution. This downloads the required number of images to cover 
a specific area at maximum zoom}
\usage{
consolidateGMMosaic(bbox = NULL, lat = NULL, lon = NULL, zoom = "max")
}
\arguments{
\item{bbox}{The bounding box. Possibly from the qbbox() function}
}
\description{
consolidateGMMosaic Function to download and stich together googleMaps 
images at a given resolution. This downloads the required number of images to cover 
a specific area at maximum zoom
}

