% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateAdditiveOutcomes.R
\name{consolodateAdditiveOutcomes}
\alias{consolodateAdditiveOutcomes}
\title{consolodateAdditiveOutcomes Function to add burden of disease outcomes in 
rasterised format based on the layer name. The expectation is that a raster 
with names like "mortality.all.cause.PM10" and "mortality.all.cause.SO2" are given 
where mortality.all.cause.PM10 is all-cause mortality due to PM10 exposure and 
mortality.all.cause.SO2 is all-cause mortality due to SO2 exposure. If the relative 
risks per pollutant used to construct the input raster represented the unique contribution
of that pollutant, the outcomes can be added. If the exposure reponse functions where 
based on co-exposure to more than one pollutant - one cannot do this or you will
have to allocate weights. Allowance must also possibly be made for synergic effects}
\usage{
consolodateAdditiveOutcomes(r, polnames = c("PM10", "SO2"))
}
\arguments{
\item{r}{A Raster with burden of disease outcomes}

\item{polnames}{A character vector with names for pollutants as they are found 
in the layernames of r}
}
\value{
A raster with one layer for each outcome
}
\description{
consolodateAdditiveOutcomes Function to add burden of disease outcomes in 
rasterised format based on the layer name. The expectation is that a raster 
with names like "mortality.all.cause.PM10" and "mortality.all.cause.SO2" are given 
where mortality.all.cause.PM10 is all-cause mortality due to PM10 exposure and 
mortality.all.cause.SO2 is all-cause mortality due to SO2 exposure. If the relative 
risks per pollutant used to construct the input raster represented the unique contribution
of that pollutant, the outcomes can be added. If the exposure reponse functions where 
based on co-exposure to more than one pollutant - one cannot do this or you will
have to allocate weights. Allowance must also possibly be made for synergic effects
}

