% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_exceed_table.R
\name{count_exceed_table}
\alias{count_exceed_table}
\title{Make table of exceedance counts}
\usage{
count_exceed_table(x, cp = "", ...)
}
\arguments{
\item{x}{a raster stack resultsing from the count_exceed function}

\item{cp}{Character The caption}
}
\description{
Make table of exceedance counts
}
\author{
Christiaan J Pauw
}

