% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diverge0.R
\name{diverge0}
\alias{diverge0}
\title{diverge0}
\usage{
diverge0(p, ramp='RdBu')
}
\arguments{
\item{p}{Trellis object generated by rasterVis::levelplot}

\item{ramp}{The name of an RColorBrewer palette (as character), a character vector of 
colour names to interpolate, or a colorRampPalette}
}
\description{
Takes a trellis object generated by rasterVis::levelplot, 
and a colour ramp, and plots the object with the colours diverging around zero.
}
\author{
J Baums
}
\references{
http://stackoverflow.com/questions/33750235/plotting-a-raster-with-the-color-ramp-diverging-around-zero
}

