% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_middel.R
\name{ext_middel}
\alias{ext_middel}
\title{Extent Middel}
\usage{
ext_middel(r, grdsize = 4, diam = 12,
  crs = "+proj=utm + south + zone=35 ellps=WGS84")
}
\arguments{
\item{r}{Raster or extent objectt}

\item{grdsize}{Numeric. The grid size.}

\item{diam}{Numeric. The diameter.}

\item{crs}{Character or object of class CRS.PROJ.4 type description of a Coordinate 
Reference System (map projection).}
}
\description{
Makes a 4km raster of a midpoint of an extent object
}

