% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop2sp.R
\name{fix.popProjek}
\alias{fix.popProjek}
\title{Fix popProjek}
\usage{
fix.popProjek(x, var = "place", asis = "maphela", mustbe = "mapehla")
}
\arguments{
\item{x}{Data frame containing the places to be fixed}

\item{var}{Character vector. The column name referring to places}

\item{asis}{Character vector. Identifier of the wronful spelling}

\item{mustbe}{Character vector to replace wrongful spelling}
}
\description{
Fixes the wrongful spelling of a place, the default case being mapehla
}

