% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eef.R
\name{get_eef}
\alias{get_eef}
\title{eef Funksie om emissie effektiwiteitsfaktor te bereken
eef_i,j = (C_ante,k,i,t,j * em.fac_pj ) / (C_post,k,i,t,j * em.fac_pj )}
\usage{
get_eef(ante = c(coal = 200, lpg = 0.001), post = c(coal = 0.01, lpg = 20),
  emf = c(coal = 12, lpg = 0.01), current = c(coal = 50, lpg = 18),
  eta = c(coal = 0.04, lpg = 0.7), NCV = c(coal = 20, lpg = 40),
  conservative = TRUE)
}
\arguments{
\item{ante}{numeric. Named vector of fuel consumption in kg before intervention}

\item{effective}{Logical. return the absolute or effective EEF}
}
\description{
eef Funksie om emissie effektiwiteitsfaktor te bereken
eef_i,j = (C_ante,k,i,t,j * em.fac_pj ) / (C_post,k,i,t,j * em.fac_pj )
}

