% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasteriseCensus.R
\name{gridzle}
\alias{gridzle}
\title{gridzle
Maak 'n rooster (SpatialGrid) van 'n raster}
\usage{
gridzle(r)
}
\arguments{
\item{r}{raster}
}
\description{
gridzle
Maak 'n rooster (SpatialGrid) van 'n raster
}

