% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactDeficit.R
\name{impdef}
\alias{impdef}
\title{impactDeficit Calculate the impact deficit the atmospheric state due to an 
emission source.}
\usage{
impdef(base, proj, kpext, hh_24, calc.prop = TRUE,
  app = "particle precursor approach")
}
\arguments{
\item{base}{A raster}

\item{proj}{A raster}
}
\value{
A raster or raster stack with the impact deficit as the difference between the baseline and
project scenario
}
\description{
impactDeficit Calculate the impact deficit the atmospheric state due to an 
emission source.
}

