% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knipNA.R
\name{knipNA}
\alias{knipNA}
\title{Cut NA}
\usage{
knipNA(r, out = c("extent", "raster")[1])
}
\arguments{
\item{r}{A raster}

\item{out}{What to return. Either "extent" for an extent obejct or "raster"}
}
\description{
Create a new extent with NA-only columns cut out
}
\references{
From an answer by Marie Auger-Methe on http://stackoverflow.com/questions/19680079/r-crop-no-data-of-a-raster
}

