% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maak_at.R
\name{maak_at}
\alias{maak_at}
\title{maak_at}
\usage{
maak_at(r, n = 20, beginnul = FALSE, maksimum = NULL, minimum = NULL,
  simetries = FALSE)
}
\arguments{
\item{r}{a Raster* objek}

\item{n}{}
}
\value{
Numeric
}
\description{
geriefsfunksie om at vir levelplot te maak
}

