% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probsel.R, R/rasteriseCensus.R
\name{make_probrast}
\alias{make_probrast}
\title{make_probrast}
\usage{
make_probrast(r)

make_probrast(r)
}
\arguments{
\item{r}{Raster}

\item{r}{Raster}
}
\description{
make_probrast

make a probability raster r/sum(r)
}

