% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_subplace.R
\name{per_subplace}
\alias{per_subplace}
\title{per_subplace Get from a raster mean per polygon}
\usage{
per_subplace(r, spdf, nm = "SP_NAME")
}
\arguments{
\item{r}{Raster}

\item{spdf}{SPDF}

\item{nm}{Character Column name in sdpf with name}
}
\description{
per_subplace Get from a raster mean per polygon
}

