% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointifyCensus.R
\name{pointifyCensus}
\alias{pointifyCensus}
\title{Pointify Census Polygon}
\usage{
pointifyCensus(spdf, dropnames = c("ID", "OBJECTID", "Geometry_s",
  "GAVPrimar0", "Geometry_1", "SP_CODE", "SP_Code", "MP_CODE", "MP_Code",
  "MN_CODE", "MN_MDB_C", "DC_MN_C", "Shape_Leng", "Shape_Area", "fakeData",
  "GAVPrimary", "Total"), verbose = TRUE, ...)
}
\arguments{
\item{spdf}{A SpatialPolygonsDataFrame}

\item{dropnames}{A character vector of column names that should not be included}

\item{verbose}{Logical. Displays function messages if TRUE}

\item{...}{Arguments to be passed to/from other methods}
}
\description{
Takes a SpatialPolygonsDataFrame with summary information, generates the 
appropriate number of random points and returns a SpatialPointsDataFrame
}
\details{
The numeric columns are counts of responses to categorical questions in the census
e.g. Which energy carrier do you use for heating.
}

