% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop2sp.R
\name{pop2sp}
\alias{pop2sp}
\title{Pop to Sp}
\usage{
pop2sp(p = fuel_house, place = "place", SPDF = KWA, var = "SP_NAME",
  raster = FALSE, refres = c(40, 40), plot = FALSE, debug = FALSE,
  resamp = FALSE, ...)
}
\arguments{
\item{p}{Data frame that contains point estimates}

\item{place}{Character vector referring to the column containing locations}

\item{SPDF}{A SpacialPolygonsDataFrame}

\item{var}{Character vector of the variable under consideration}

\item{raster}{Logical to Rasterise a census polygon SPDF possibly containing 
summary data if TRUE.}

\item{refres}{Numeric vector of length 2 (x,y): A reference resolution}

\item{plot}{Logical that creates a level and contour plot if TRUE}

\item{debug}{Logical to display values for debugging process}

\item{...}{Arguments to be passed to/from other methods}
}
\description{
Function that edits and plots a SpacialPolygonsDataFrame
}

