% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_exceed.R
\name{print_exceed}
\alias{print_exceed}
\title{Print Exceed}
\usage{
print_exceed(z, type = c("k", "l")[1], ttl = "Count of exceedences")
}
\arguments{
\item{z}{A raster or raster brick}

\item{type}{Character vector that contains the desired output. k = kable, l = LaTeX.}

\item{ttl}{Character vector that contains the title}
}
\description{
Print statistics for the cells of each layer of a raster object
}

