% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probsel.R
\name{probsel}
\alias{probsel}
\title{spatial sampling, with selection probabilities defined in a raster}
\source{
https://scrogster.wordpress.com/2012/04/22/using-r-for-spatial-sampling-with-selection-probabilities-defined-in-a-raster/
}
\usage{
samppoints<-probsel(probrast, 300)
}
\arguments{
\item{probrast}{Raster with probabilities sum(getValues(probrast), na.rm=T) should be 1}

\item{N}{integer Number of points to select}
}
\value{
SpatialPoints
}
\description{
spatial sampling, with selection probabilities defined in a raster
}
\author{
Michael Scroggie
}

