% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_population_brick.R
\name{proportion_sex}
\alias{proportion_sex}
\title{make_population_brick Make a raster brick that is the equavalent of a population pyramid}
\usage{
proportion_sex(r, prop.men = 0.48)
}
\arguments{
\item{r}{A Raster containing population numbers for males and females. If you only have the total, proportion_sex() will 
be used internally to make a two layer raster brick with male and female numbers. You can provide a prop.men if you have it 
(else the default of 0.48 will be used)}

\item{prop}{A numeric vector giving the proportion of the population per age or sex group}

\item{DF}{a data frame containing A column of age groups cut-offs, a column of proportion of mae and a column of proportion of females}

\item{groupname}{The variable name in DF that contains the age group descriptions}

\item{malename}{The variable name in DF that contains proportions of males per age group}

\item{femalename}{The variable name in DF that contains proportions of females per age group}
}
\description{
make_population_brick Make a raster brick that is the equavalent of a population pyramid
}
\details{
Every layer represents an age sex-group e.g. men-42 represents men aged 
older or equal to 42 and less than 43. The use of this is to apply an CRF to each 
layer depending on whether that layer qualifies. E.g. If the function is defined for 
people above 18 then a function must select only the layers that meet the requirement
}

