% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{PSI}
\alias{PSI}
\title{Pollutant Sub Index}
\usage{
PSI(x, Exp.met, verbose = FALSE)
}
\arguments{
\item{x}{A raster}

\item{Exp.met}{Coefficients to be considered}

\item{verbose}{Logical to display function messages if TRUE}
}
\description{
Function that calculates and returns the PSI
}

