% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_dist_plot.R
\name{raster_dist_plot}
\alias{raster_dist_plot}
\title{raster_dist_plot Convenience function to summarise a raster stack and plot the result}
\usage{
raster_dist_plot(ss, multi = list(NULL, c("IndustryY_pm10",
  "IndustryX_so2"))[[1]], mn = "", sb = "", th = BuRdTheme,
  scales = list(draw = FALSE), ...)
}
\arguments{
\item{ss}{A raster stack}

\item{multi}{A character vector of layer name components or NULL}

\item{mn}{The main title of the resulting plot as a character string. Defaults to empty string}

\item{sb}{The plot subtitle as a character string}
}
\value{
A trellis plot
}
\description{
raster_dist_plot Convenience function to summarise a raster stack and plot the result
}

