% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{rasterAPI}
\alias{rasterAPI}
\title{Raster Air Pollution Index}
\usage{
rasterAPI(s, reftab = RR.accut.mort, idpos = 1, polpos = 2,
  aveperiodpos = 3, cyclepos = 5, sep = "_", aggregate = FALSE,
  per_source = FALSE, verbose = FALSE, return.full = FALSE)
}
\arguments{
\item{s}{Raster or a raster stack}

\item{reftab}{Data frame containing information relating to accute mortality}

\item{idpos}{Numeric referring to the ID position}

\item{polpos}{Numeric referring to the position of the pollutant.}

\item{aveperiodpos}{Numeric referring to the average period position.}

\item{cyclepos}{Numeric referring to the cycle position.}

\item{sep}{Character vector containing the seperator to be used.}

\item{aggregate}{Logical that initializes aggregation when TRUE.}

\item{per_source}{Logical that returns the source ID.}

\item{verbose}{Logical that displays function messages if TRUE.}

\item{return.full}{Logical that stacks the raster objects and returns it when TRUE.}
}
\description{
Creates a raster for the API
}

