% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterHour2Day.R
\name{rasterH2D}
\alias{rasterH2D}
\title{Hour to day}
\usage{
rasterH2D(s, knip = FALSE)
}
\arguments{
\item{s}{A raster}

\item{knip}{Logical. Create a new extent with NA-only columns cut out if TRUE}
}
\description{
Converts an hour raster to a day raster
}

