% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWI.R
\name{rasterSWI}
\alias{rasterSWI}
\title{Raster SWI}
\usage{
rasterSWI(s, reftab = NAQS, idpos = 1, polpos = 2, aveperiodpos = 3,
  cyclepos = 5, sep = "_", verbose = TRUE)
}
\arguments{
\item{s}{A raster or raster stack}

\item{reftab}{A data frame with the standards. 
Colnames are the averaging periods. Rownames are the pollutants.}

\item{idpos}{Numeric. Position of the ID column}

\item{polpos}{Numeric. Position of the pollutant column}

\item{aveperiodpos}{Numeric. Position of the column containing everage periods}

\item{cyclepos}{Numeric. Position of the cycle}

\item{sep}{Character vector containing the seperator to be used.}

\item{verbose}{Logical that displays function messages if TRUE.}
}
\description{
Creates a raster for the standars weighted intake
}

