% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.select.cdf.R
\name{read.select.cdf}
\alias{read.select.cdf}
\title{Read Select .cdf}
\usage{
read.select.cdf(fn = file.choose(), select = NULL, verbose = FALSE)
}
\arguments{
\item{fn}{Character vector. The name of the raster file. Defaults to file.choose()}

\item{select}{Regular expression. A part of the name that you want to use to select layers. 
NULL to select all layers. Defaults to NULL}

\item{verbose}{Do you want additional messages.}
}
\value{
A Raster* object
}
\description{
Convenience function to read cdf file and select layers
}

