% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/End.Point.Info.China.R
\name{rebase.rr}
\alias{rebase.rr}
\title{Rebase RR}
\usage{
rebase.rr(RR, base.orig = 10)
}
\arguments{
\item{RR}{A raster}

\item{base.orig}{Numeric. The original base concentration}
}
\description{
Modifies the base concentrtion of a raster
}

