% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\docType{data}
\name{RR.accut.mort}
\alias{RR.accut.mort}
\title{Air Pollution Index}
\format{An object of class \code{data.frame} with 6 rows and 7 columns.}
\usage{
RR.accut.mort
}
\description{
Air Pollution Index
}
\references{
Air Quality Index from Cairncross, John and Zunckel Atm.Envir 41 (2007)
}
\keyword{datasets}

