% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumIncidences.R
\name{sumIncidences}
\alias{sumIncidences}
\title{sumIncidences}
\usage{
sumIncidences(outcomesStack)
}
\arguments{
\item{outcomesStack}{Raster stack. assumes that outcomesStack is a raster stack with 1 layer per outcome}
}
\description{
Sommeer die uitkoms van aggregateCREPstack
}

