% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.sicklist.R
\name{summarise.sicklist}
\alias{summarise.sicklist}
\title{Summarise Sicklist}
\usage{
summarise.sicklist(sicklist)
}
\arguments{
\item{sicklist}{A sicklist}
}
\description{
Script to summarise a sicklist into dataframe for presentation in report
}

