% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_PTI.R
\name{summary_PTI}
\alias{summary_PTI}
\title{Summary_PTI}
\usage{
summary_PTI(s, msk = KWA[which(KWA$SP_NAME == "Kwazamokuhle SP"), ],
  nm = NULL)
}
\arguments{
\item{s}{RasterStack}

\item{msk}{SPDF to act as mask or NULL}

\item{nm}{Character Name of row or NULL}
}
\description{
Summarise the proportion to implement. Convenience function to
  print table of proportion to implement. Possibly masking with a SPDF
}

