% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWI.R
\name{SWI}
\alias{SWI}
\title{Standards Weighted Intake}
\usage{
SWI(conc, NAQS = NULL, pop = pop, ap = "Annual", Q = 8,
  refpol = c("PM10", "PM2.5", "SO2", "O3", "O3.1", "NO2")[1])
}
\arguments{
\item{conc}{raster The concentration of every pollutant}

\item{pop}{raster The # of people in each grid cell}

\item{ap}{Character "Annual" or "Daily"}

\item{Q}{numeric Breathing rate. Default 8 m3 per day}

\item{refpol}{one of c("PM10", "PM2.5", "SO2", "O3", "O3.1", "NO2"). Default is "PM10"}

\item{reftab}{A data frame with the standards. Colnames are the averaging periods. Rownames are the pollutants}
}
\description{
Calculates the standars weighter intake of a pollutant
}

