% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_exceed.R
\name{table_exceed}
\alias{table_exceed}
\title{table_exceed Tabulation fucntion to accompany count_exceed(). Show the number 
of cells exceeding each level the raster or per subplace. e.g 
Level Suburb Exceedence_Count}
\usage{
table_exceed()
}
\description{
table_exceed Tabulation fucntion to accompany count_exceed(). Show the number 
of cells exceeding each level the raster or per subplace. e.g 
Level Suburb Exceedence_Count
}

