% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knipNA.R
\name{knipNA}
\alias{knipNA}
\title{knipNA create a new extent with NA-only columns cut out}
\usage{
knipNA(r, out = c("extent", "raster")[1])
}
\description{
knipNA create a new extent with NA-only columns cut out
}

