% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_ratio.R
\name{particle_ratio}
\alias{particle_ratio}
\title{particle_ratio Calculate the equavalence of sources in terms of the 
resultant concentrations at each receptor}
\usage{
particle_ratio(source1 = rep(400, 365),
  conc1 = baseline365[[grep("Indus[[:print:]]*_pm10[[:print:]]",
  names(baseline365))]], source2 = rep(x = 15, 365),
  conc2 = baseline365[[grep("H[[:print:]]*_pm10[[:print:]]*",
  names(baseline365))]])
}
\arguments{
\item{source1}{numeric vector of the same length as nlayers(conc1). Total emissions}

\item{conc1}{A Raster object with concentrations resulting from source1.}

\item{source2}{numeric vector of the same length as nlayers(conc2)}

\item{conc2}{A Raster object with concentrations resulting from source2}
}
\value{
A raster or raster stack with the source equavalence ratios (source impact ratio)
. A large number means there is a large offset opportunity or that the offset source (source 2) has a
much larger impact than the source to be offset (source1)
}
\description{
particle_ratio Calculate the equavalence of sources in terms of the 
resultant concentrations at each receptor
}
\author{
Rebecca Garland, Mogesh Naidoo, Christiaan Pauw
}

