% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_dist_sum.R
\name{raster_ave_type}
\alias{raster_ave_type}
\title{Raster Average Type}
\usage{
raster_ave_type(s, patt = "_24h\\\\.[[:digit:]]+", func = "mean",
  verbose = FALSE, ...)
}
\arguments{
\item{s}{A raster stack}

\item{patt}{Character vector used to isolate unique classes within s . typilcally the varying part. e.g
if the layers are names pm10_1, pm10_2 ....the patt is '_[[:digit:]]+$"}

\item{func}{The function to be applied on subsets of the raster stack}

\item{verbose}{Logical Messages or not}
}
\value{
Raster with function applied to subsets
}
\description{
Calculates the averages of unique names within a raster stack
}

