% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_dist_plot.R
\name{raster_dist_plot}
\alias{raster_dist_plot}
\title{Raster Distribution plot}
\usage{
raster_dist_plot(ss, multi = list(NULL, c("IndustryY_pm10",
  "IndustryX_so2"))[[1]], mn = "", sb = "", th = BuRdTheme,
  scales = list(draw = FALSE), meanonly = FALSE, tab_out = FALSE,
  debug = FALSE, verbose = FALSE, beginnul = FALSE, maksimum = NULL,
  diverge = FALSE, ...)
}
\arguments{
\item{ss}{A raster stack}

\item{multi}{A character vector of layer name components or NULL}

\item{mn}{The main title of the resulting plot as a character string. Defaults to empty string}

\item{sb}{The plot subtitle as a character string}

\item{th}{The theme to be applied temporarily to the plotting procedure}

\item{scales}{A list with components x (columns) and     y (rows). Each of 
these components must be a numeric vector of length 2 defining the range for each marginal plot.
Default contains a logical that prevents drawing marginal graphics}

\item{meanonly}{Logical Return only the mean}

\item{tab_out}{Logical Write out a table}

\item{...}{Arguments to be passed to raster_dist_sum}
}
\description{
Convenience function to summarise a raster stack and plot the result as a 
trellis plot summarising the distribution by quantiles
}

