% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasteriseCensus.R
\name{sexify}
\alias{sexify}
\title{Sexify}
\usage{
sexify(x, prop = 0.5)
}
\arguments{
\item{x}{A SpatialPolygonsDataFrame}

\item{prop}{Numeric that contains the sex proportion}
}
\description{
Change  SPDF to match a se proportion
}

