% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWI.R
\name{SWI}
\alias{SWI}
\title{Standards Weighted Intake}
\usage{
SWI(conc, NAQS = NULL, pop = people, ap = "Annual", Q = 8)
}
\arguments{
\item{conc}{Raster. The concentration of every pollutant}

\item{NAQS}{Data frame of the National Air Quality Standards}

\item{pop}{Raster. The number of people in each grid cell}

\item{ap}{Character verctor. "Annual" or "Daily"}

\item{Q}{Numeric. Breathing rate in m3 per person per day}
}
\description{
Calculates the total effective intake of pollutants by using
standardisation
}

