% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWI.R
\name{SWI}
\alias{SWI}
\title{Standards Weighted Intake}
\usage{
SWI(conc, NAQS = NULL, pop = people, ap = "Annual", Q = 8, idpos = 1,
  polpos = 3, aveperiodpos = 2, cyclepos = 5, sep = "_",
  verbose = FALSE)
}
\arguments{
\item{conc}{Raster. The concentration of every pollutant}

\item{NAQS}{Data frame of the National Air Quality Standards}

\item{pop}{Raster. The number of people in each grid cell}

\item{ap}{Character verctor. "Annual" or "Daily"}

\item{Q}{Numeric. Breathing rate in m3 per person per day}

\item{idpos}{Numeric referring to the ID position}

\item{polpos}{Numeric referring to the position of the pollutant.}

\item{aveperiodpos}{Numeric referring to the average period position.}

\item{cyclepos}{Numeric referring to the cycle position.}

\item{sep}{Character vector containing the seperator to be used.}
}
\description{
Calculates the total effective intake of pollutants by using
standardisation
}

